+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/auth/FindPasswordActivity.kt
@@ -1,16 +1,14 @@
-package com.wonchihyeon.ytt_android.account
+package com.wonchihyeon.ytt_android.auth
 
 import android.os.Bundle
 import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.wonchihyeon.ytt_android.R
 
-class FindIdActivity : AppCompatActivity() {
+class FindPasswordActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-        setContentView(R.layout.activity_find_id)
+        setContentView(R.layout.activity_find_password)
     }
 }
