+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/auth/LoginActivity.kt
@@ -1,4 +1,4 @@
-package com.wonchihyeon.ytt_android.account
+package com.wonchihyeon.ytt_android.auth
 
 import android.content.Intent
 import android.content.pm.PackageInfo
