+++ FE/app/src/main/java/com/wonchihyeon/ytt_android/order/OrderItem.kt
@@ -1,4 +1,4 @@
-package com.wonchihyeon.ytt_android
+package com.wonchihyeon.ytt_android.order
 
 data class OrderItem(val name: String, val basePrice: String, val totalPrice: String)
